requirement.txt
asgiref                       
Django                       
djangorestframework           
djangorestframework_simplejwt 
pip                          
PyJWT                         
setuptools                    
sqlparse                      
tzdata                       